실습1

+++ main.c
@@ -5,14 +5,14 @@
 
 int main(int argc, char *argv[]) {
 	
-	int number;
+	double number;
 	printf(" 정수 하나를 입력하시오 :");
-	scanf("%i", &number);
+	scanf("%d", &number);
 	 
-	 if("%i>0", number )
+	 if("%d>0", number )
 	 printf(" 양수입니다.");
 	 
-	 else if("%i<0", number)
+	 else if("%d<0", number)
 	 printf(" 음수입니다.");
 	 
 	 else
